Fix hook handling
Setting the hook for the duration of calling the composefunc of the
fallback mail user agent won't work as intended, because the hook is
called when mail is _sent_, not when the compose buffer is opened.

As far as I can tell, there's nothing we can easily do to install a
hook for our fallback MUA; the best we can do is to show an
informative error message.